refactor to be independent from junit and spring
