+++ spring-cloud-gcp-data-spanner/src/test/java/org/springframework/cloud/gcp/data/spanner/test/IntegrationTestConfiguration.java
@@ -26,6 +26,7 @@ import com.google.cloud.spanner.Spanner;
 import com.google.cloud.spanner.SpannerOptions;
 
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.cloud.gcp.core.Credentials;
 import org.springframework.cloud.gcp.core.DefaultCredentialsProvider;
 import org.springframework.cloud.gcp.core.DefaultGcpProjectIdProvider;
@@ -104,6 +105,7 @@ public class IntegrationTestConfiguration {
 	}
 
 	@Bean
+	@ConditionalOnMissingBean
 	public SpannerOptions spannerOptions() {
 		return SpannerOptions.newBuilder().setProjectId(getProjectId())
 				.setSessionPoolOption(SessionPoolOptions.newBuilder().setMaxSessions(10).build())
@@ -116,6 +118,7 @@ public class IntegrationTestConfiguration {
 	}
 
 	@Bean
+	@ConditionalOnMissingBean
 	public Spanner spanner(SpannerOptions spannerOptions) {
 		return spannerOptions.getService();
 	}
