+++ spring-cloud-gcp-test-support/src/main/java/org/springframework/cloud/gcp/test/PubSubEmulatorRule.java
@@ -16,39 +16,29 @@
 
 package org.springframework.cloud.gcp.test;
 
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
+import org.junit.rules.ExternalResource;
 
-public class PubSubEmulatorRule extends GcpEmulatorRule {
-	private static final Log LOGGER = LogFactory.getLog(PubSubEmulatorRule.class);
+public class PubSubEmulatorRule extends ExternalResource {
 
-  String getGatingPropertyName() {
-		return "it.pubsub-emulator";
-	}
+  private PubSubEmulatorHelper emulatorHelper = new PubSubEmulatorHelper();
 
-	String getEmulatorName() {
-		return "pubsub";
+	@Override
+	protected void before() throws Throwable {
+		emulatorHelper.startEmulator();
 	}
 
 	@Override
-	protected void afterEmulatorDestroyed() {
-		String hostPort = getEmulatorHostPort();
-
-		// find destory emulator process spawned by gcloud
-		if (hostPort == null) {
-			LOGGER.warn("Host/port null after the test.");
-		} else {
-			int portSeparatorIndex = hostPort.lastIndexOf(":");
-			if (portSeparatorIndex < 0) {
-				LOGGER.warn("Malformed host: " + hostPort);
-				return;
-			}
-
-			String emulatorHost = hostPort.substring(0, portSeparatorIndex);
-			String emulatorPort = hostPort.substring(portSeparatorIndex + 1);
-
-			String hostPortParams = String.format("--host=%s --port=%s", emulatorHost, emulatorPort);
-			killByCommand(hostPortParams);
-		}
+	protected void after() {
+		emulatorHelper.shutdownEmulator();
 	}
+
+	/**
+	 * Return the already-started emulator's host/port combination when called from within a
+	 * JUnit method.
+	 * @return emulator host/port string or null if emulator setup failed.
+	 */
+	public String getEmulatorHostPort() {
+		return emulatorHelper.getEmulatorHostPort();
+	}
+
 }
