+++ spring-cloud-gcp-test-support/src/main/java/org/springframework/cloud/gcp/test/SpannerEmulatorRule.java
@@ -16,51 +16,28 @@
 
 package org.springframework.cloud.gcp.test;
 
-import java.util.stream.Collectors;
 import org.junit.Assume;
+import org.junit.rules.ExternalResource;
 
 /**
  * TODO: send users to https://cloud.google.com/spanner/docs/emulator#using_the_gcloud_cli_with_the_emulator
  */
-public class SpannerEmulatorRule extends GcpEmulatorRule {
-	String getGatingPropertyName() {
-		return "it.spanner-emulator";
-	}
+public class SpannerEmulatorRule extends ExternalResource {
 
-	String getEmulatorName() {
-		return "spanner";
-	}
+	private SpannerEmulatorHelper emulatorHelper = new SpannerEmulatorHelper();
 
 	@Override
-	protected void beforeEmulatorStart() {
+	protected void before() throws Throwable {
 		String emulatorHost = System.getenv("SPANNER_EMULATOR_HOST");
 		Assume.assumeFalse(
-				"Run this command prior to running an emulator test:\n$(gcloud beta emulators spanner env-init)",
+				"Run this command prior to running an emulator test and set SPANNER_EMULATOR_HOST environment variable:\ngcloud beta emulators spanner env-init",
 				emulatorHost == null || emulatorHost.isEmpty());
+		emulatorHelper.startEmulator();
 	}
 
 	@Override
-	protected void afterEmulatorStart() {
-		ProcessOutcome switchToEmulator = runSystemCommand(new String[] {
-				"gcloud", "config", "configurations", "activate", "emulator"});
-
-		ProcessOutcome processOutcome = runSystemCommand(new String[] {
-				"gcloud", "spanner", "instances", "create", "integration-instance", "--config=emulator-config", "--description=\"Test Instance\"", "--nodes=1" },
-				false);
-
-		if (processOutcome.getStatus() != 0) {
-			// don't set breakpoint here
-			this.killByCommand("cloud_spanner_emulator/emulator_main");
-			throw new RuntimeException("Creating instance failed: "
-					+ String.join("\n", processOutcome.getErrors()));
-		}
-
-
-		// TODO: don't forget to kill the 2 spanner processes
+	protected void after() {
+		emulatorHelper.shutdownEmulator();
 	}
 
-	/*
-	* gcloud spanner instances create test-instance    --config=emulator-config --description="Test Instance" --nodes=1
-	* gcloud config configurations activate [emulator | default]
-	* */
 }
